The previous hibernate/jpa configuration was working but there was a conflicting hibernate (spring xml) config going against persistence.xml (jpa) config. The hibernate (spring xml) config wasn't being picked up anyway. Unnecessary config was removed. JPA config is now added. New EntityManager is built.

+++ billrive-service/src/main/java/com/uhsarp/billrive/domain/UserIdAndLiableCost.java
@@ -15,7 +15,7 @@ import javax.persistence.*;
 public class UserIdAndLiableCost  implements GenericObject {
     
     private  int userId;
-    private  int liableCost;
+    private  float liableCost;
     private int billItemEntryID;
 
     public UserIdAndLiableCost() {
@@ -42,7 +42,7 @@ public class UserIdAndLiableCost  implements GenericObject {
         this.userId = userId;
     }
 
-    public int getLiableCost() {
+    public float getLiableCost() {
         return liableCost;
     }
